New application packages structure, refactoring
