--- application/interface_message_handler.go
@@ -1,7 +1,0 @@
-package application
-
-type MessageHandlerInterface interface {
-	ValidateAndParse(message string) (uint64, error)
-	Handle(number uint64)
-	Terminate()
-}
