+++ cmd/handler/message_handler_interface.go
@@ -1,0 +1,6 @@
+package handler
+
+type MessageHandlerInterface interface {
+	ValidateAndParse(message string) (uint64, error)
+	Handle(number uint64)
+}
