--- domain/interface_dumper.go
@@ -1,7 +1,0 @@
-package domain
-
-import "bufio"
-
-type DumperInterface interface {
-	ProcessChannel(inChannel chan uint64, writer *bufio.Writer)
-}
