--- domain/number_storage_interface.go
@@ -1,7 +1,0 @@
-package domain
-
-type NumberStorageInterface interface {
-	IsNumberExists(number uint64) bool
-	AddNumber(number uint64) (bool, error)
-	GetLength() uint64
-}
