+++ internal/storage/number_storage_interface.go
@@ -1,0 +1,7 @@
+package storage
+
+type NumberStorageInterface interface {
+	IsNumberExists(number uint64) bool
+	AddNumber(number uint64) (bool, error)
+	GetLength() uint64
+}
